cortex-backend/requirement.txt
fastapi[standard]
tortoise-orm[asyncpg]
aerich
python-dotenv
passlib==1.7.4
bcrypt==4.0.1
pyjwt